% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpuscleaner-package.R
\docType{package}
\name{corpuscleaner-package}
\alias{corpuscleaner}
\alias{corpuscleaner-package}
\title{corpuscleaner: A Tibble of 18 Cleaned Documents usable for Text Analysis}
\description{
The package has two major functions novels() and polish(..., vars). To know more about them please refer to the Vignette.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://anshalmm.github.io/corpuscleaner/}
}

}
\author{
\strong{Maintainer}: Anshal Majmudar \email{ammajmudar@ucdavis.edu}

}
\keyword{internal}
