% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_by_ID.R
\name{document_by_ID}
\alias{document_by_ID}
\title{Documents Extraction}
\usage{
document_by_ID(...)
}
\arguments{
\item{...}{Input any ID (i.e., 1:18) from the `Documents` dataset}

\item{vars}{Input any column variable from the `Documents` dataset}
}
\value{
A \code{tibble} with `id`, and then any column variable of your choice
}
\description{
This function allows you to extract the document of your choice based on the title you choose along with
any extra parameters you'd like to see with it such as the `year` or the `author`.
}
\examples{
document_by_ID(id == 11, vars = c("year", "author"))
}
