% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels_doc.R
\docType{data}
\name{Documents}
\alias{Documents}
\title{Dataset of 18 Cleaned Documents}
\format{
A \code{tibble} with 369111 rows and 5 variables consisting of:
\describe{
 \item{id}{ID of Documents}
 \item{title}{Title of Documents}
 \item{text}{Text of Documents}
 \item{year}{Year Published}
 \item{authors}{Authors of Documents}
}
}
\usage{
Documents
}
\description{
A Tibble of 18 Cleaned Documents usable for Corpus Analysis
}
\keyword{dataset}
