% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\name{novels}
\alias{novels}
\title{18 Cleaned Novels}
\usage{
novels()
}
\value{
A dataframe with 369111 rows and 5 variables
}
\description{
A complete cleaned tibble of 18 Novels that is usable for Corpus Analytics
}
\examples{
library(dplyr)
library(tidytext)
novels() \%>\%
  unnest_tokens(word, text) \%>\%
  anti_join(get_stopwords("en", source = "smart")) \%>\%
  count(ID, title, word, sort = T)
}
