% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels_by_ID.R
\name{novels_by_ID}
\alias{novels_by_ID}
\title{Novels Extraction}
\usage{
novels_by_ID(ID, vars)
}
\arguments{
\item{ID}{Input any ID (i.e., 1:18) from the novels() function}

\item{vars}{Input any column variable of the novels() function}
}
\value{
A dataframe with ID, text, and then any column variable of your choice
}
\description{
This function allows you to extract the novel(s) of your choice based on the title you choose along with
any extra parameters you'd like to see with it such as the year or the author.
}
\examples{
novels_by_ID(ID = c(5, 6), vars = "year")
}
