% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels_doc.R
\docType{data}
\name{novels_tgt}
\alias{novels_tgt}
\title{A Tibble of 18 Cleaned Novels}
\format{
A \code{tibble} with 369111 rows and 5 variables:
\describe{
 \item{title}{Title of Novels}
 \item{text}{Text of Novels}
 \item{year}{Year Published}
 \item{author}{Author of Novels}
 \item{novel_bookshelf}{Category of Novels}
 }
}
\usage{
novels_tgt
}
\description{
A \code{tibble} of 18 Cleaned Novels that is usable for Corpus Analysis. It consists of
}
\keyword{dataset}
