% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish.R
\name{polish}
\alias{polish}
\title{Individual Novels Extraction}
\usage{
polish(..., vars)
}
\arguments{
\item{...}{Input any title from the novels() function}

\item{vars}{Input any column variable of the novels() function}
}
\value{
A dataframe with title, text, and then any column variable of your choice
}
\description{
This function allows you to extract the novel of your choice based on the title you choose along with
any extra parameters you'd like to see with it such as the year or the author.
}
\examples{
polish(title == "Frankenstein", vars = "year")
}
