% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_token.R
\name{unnest_bigrams}
\alias{unnest_bigrams}
\title{unnest_bigrams}
\format{
A \code{tibble} with 5 variables:
\describe{
 \item{id}{ID of the Document}
 \item{title}{Title of the Document}
 \item{author}{Author of the Document}
 \item{year}{Year Published}
 \item{token_two}{Token Column}
}
}
\usage{
unnest_bigrams(data, args, ...)
}
\arguments{
\item{...}{Input in the `id` of your choice from the `Documents` dataset}
}
\value{
Mulitword combination of `2 tokens`
}
\description{
Tokenizing the text into 2 units (Words)
}
\examples{
BDocument = document_by_ID(id == 5)
unnest_bigrams(BDocument)
}
