% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_token.R
\name{unnest_N4_grams}
\alias{unnest_N4_grams}
\title{unnest_N4_grams}
\format{
A \code{tibble} with 5 variables:
\describe{
 \item{id}{ID of the Document}
 \item{title}{Title of the Document}
 \item{author}{Author of the Document}
 \item{year}{Year Published}
 \item{token_four}{Token Column}
}
}
\usage{
unnest_N4_grams(data, args)
}
\arguments{
\item{tbl}{Input in the `tibble` of your choice made from the `document_by_ID` function}
}
\value{
Mulitword combination of `4 tokens`
}
\description{
Tokenizing the text into 4 units (Words)
}
\examples{
NDocument = document_by_ID(id == 10)
unnest_N4_grams(NDocument)
}
