% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_token.R
\name{unnest_trigrams}
\alias{unnest_trigrams}
\title{unnest_trigrams}
\format{
A \code{tibble} with 5 variables:
\describe{
 \item{id}{ID of the Document}
 \item{title}{Title of the Document}
 \item{author}{Author of the Document}
 \item{year}{Year Published}
 \item{token_three}{Token Column}
}
}
\usage{
unnest_trigrams(data, args)
}
\arguments{
\item{tbl}{Input in the `tibble` of your choice made from the `document_by_ID` function}
}
\value{
Mulitword combination of `3 tokens`
}
\description{
Tokenizing the text into 3 units (Words)
}
\examples{
TDocument = document_by_ID(id == 7)
unnest_trigrams(TDocument)
}
