% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels.R
\name{novels}
\alias{novels}
\title{18 Cleaned Novels}
\format{
A dataframe with 369111 rows and 6 variables consisting of:
\describe{
 \item{ID}{ID of Documents}
 \item{title}{Title of Documents}
 \item{text}{Text of Documents}
 \item{year}{Year Published}
 \item{author}{Author of Documents}
}
}
\usage{
novels()
}
\description{
A complete cleaned tibble of 18 Documents that are usable for Corpus Analytics
}
\examples{
library(dplyr)
library(tidytext)
novels() \%>\%
  unnest_tokens(word, text) \%>\%
  anti_join(get_stopwords("en", source = "smart")) \%>\%
  count(ID, title, word, sort = TRUE)
}
