% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels_doc_two.R
\docType{data}
\name{novels_tgt}
\alias{novels_tgt}
\title{A Tibble of 18 Cleaned Documents usable for Corpus Analysis with one row per line of text.}
\format{
A \code{tibble} with 369111 rows and 5 variables consisting of:
\describe{
 \item{ID}{ID of Documents}
 \item{title}{Title of Documents}
 \item{text}{Text of Documents}
 \item{year}{Year Published}
 \item{authors}{Authors of Documents}
}
}
\usage{
novels_tgt
}
\description{
A Tibble of 18 Cleaned Documents usable for Corpus Analysis with one row per line of text.
}
\keyword{dataset}
