% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/novels_doc.R
\docType{data}
\name{novels_tgt}
\alias{novels_tgt}
\title{18 Cleaned Documents}
\format{
A \code{tibble} with 369111 rows and 5 variables consisting of:
\describe{
 \item{ID}{ID of Novels}
 \item{title}{Title of Novels}
 \item{text}{Text of Novels}
 \item{year}{Year Published}
 \item{author}{Author of Novels}
}
}
\usage{
novels_tgt
}
\description{
A Tibble of 18 Cleaned Documents usable for Corpus Analysis
}
\keyword{dataset}
