% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_token.R
\name{unnest_N5_grams}
\alias{unnest_N5_grams}
\title{unnest_N5_grams}
\format{
A \code{tibble} with 5 variables:
\describe{
 \item{id}{ID of the Document}
 \item{title}{Title of the Document}
 \item{author}{Author of the Document}
 \item{year}{Year Published}
 \item{token_five}{Token Column}
}
}
\usage{
unnest_N5_grams(...)
}
\arguments{
\item{...}{Input in the `id` of your choice from the `Documents` dataset}
}
\value{
Mulitword combination of `5 tokens`
}
\description{
Tokenizing the text into 5 units (Words)
}
\examples{
unnest_N5_grams(id == 2)
}
