% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_token.R
\name{unnest_unigrams}
\alias{unnest_unigrams}
\title{unnest_unigrams}
\format{
A \code{tibble} with 5 variables:
\describe{
 \item{id}{ID of the Document}
 \item{title}{Title of the Document}
 \item{author}{Author of the Document}
 \item{year}{Year Published}
 \item{word}{Token Column}
}
}
\usage{
unnest_unigrams(data, args, ...)
}
\arguments{
\item{...}{Input in the `id` of your choice from the `Documents` dataset}
}
\value{
Splitting text into `1 token`
}
\description{
Tokenizing the text into 1 unit (Word)
}
\examples{
UDocument = document_by_ID(id == 1)
unnest_unigrams(UDocument)
}
